+++ kaiax/noop/impl/execution.go
@@ -14,15 +14,22 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with the Kaia library. If not, see <http://www.gnu.org/licenses/>.
 
-package types
+package impl
 
-import "github.com/kaiachain/kaia/kaiax"
+import (
+	"github.com/kaiachain/kaia/blockchain/types"
+	"github.com/kaiachain/kaia/common"
+)
 
-type NoopModule interface {
-	kaiax.BaseModule
-	kaiax.JsonRpcModule
-	kaiax.ConsensusModule
-	kaiax.ExecutionModule
-	kaiax.RewindableModule
-	kaiax.TxProcessModule
+func (m *NoopModule) PostInsertBlock(block *types.Block) error {
+	logger.Info("NoopModule PostInsertBlock", "blockNum", block.Header().Number.Uint64())
+	return nil
+}
+
+func (m *NoopModule) RewindTo(block *types.Block) {
+	logger.Info("NoopModule RewindTo", "blockNum", block.Header().Number.Uint64())
+}
+
+func (m *NoopModule) RewindDelete(hash common.Hash, num uint64) {
+	logger.Info("NoopModule RewindDelete", "num", num)
 }
