+++ kaiax/noop/impl/txprocess.go
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with the Kaia library. If not, see <http://www.gnu.org/licenses/>.
 
-package noop
+package impl
 
 import (
 	"github.com/kaiachain/kaia/blockchain/types"
@@ -22,11 +22,11 @@ import (
 )
 
 func (m *NoopModule) PreRunTx(evm *vm.EVM, tx *types.Transaction) (*types.Transaction, error) {
-	logger.Info("NoopModule PreRunTx")
+	logger.Info("NoopModule PreRunTx", "tx", tx.Hash())
 	return tx, nil
 }
 
 func (m *NoopModule) PostRunTx(evm *vm.EVM, tx *types.Transaction) error {
-	logger.Info("NoopModule PostRunTx")
+	logger.Info("NoopModule PostRunTx", "tx", tx.Hash())
 	return nil
 }
