+++ kaiax/noop/interface.go
@@ -16,20 +16,13 @@
 
 package noop
 
-import (
-	"github.com/kaiachain/kaia/blockchain/types"
-	"github.com/kaiachain/kaia/common"
-)
+import "github.com/kaiachain/kaia/kaiax"
 
-func (m *NoopModule) PostInsertBlock(block *types.Block) error {
-	logger.Info("NoopModule PostInsertBlock", "blockNum", block.Header().Number.Uint64())
-	return nil
-}
-
-func (m *NoopModule) RewindTo(block *types.Block) {
-	logger.Info("NoopModule RewindTo", "blockNum", block.Header().Number.Uint64())
-}
-
-func (m *NoopModule) RewindDelete(hash common.Hash, num uint64) {
-	logger.Info("NoopModule RewindDelete", "num", num)
+type NoopModule interface {
+	kaiax.BaseModule
+	kaiax.JsonRpcModule
+	kaiax.ConsensusModule
+	kaiax.ExecutionModule
+	kaiax.RewindableModule
+	kaiax.TxProcessModule
 }
